Introduces Cache Control Config Key
This adds a configuration key of GHOST_STORAGE_ADAPTER_R2_CACHE_CONTROL which takes a string value.

If unset, it does not save the file with a CacheControl header. This way it will default to the buckets configuration when unset.

An example value would be "public, max-age=2592000" which would instruct to cache the resource for 30 days but CRUCIALLY by setting "public" we can set the bucket to private and require Authorization headers but still cache the resource.